Attempted to get furniture drawing
Messing around trying get furniture to draw on the drawingpanel. Logic seems mostly right. I add the button and icon to a hash map, and then store the last clicked icon. Try to draw last clicked icon on screen. The commented out section in DrawingPanel was to address the bug where temp segments where not being  redrawn properly(displaying multiple lines), the code does fix it. It also introduces the new problem which is erasing already drawn lines till the button is let go and the whole panel is repainted.