Furniture code integrated
There is still a problem with the logic. I think something with the mouse action listeners. It loads the toolbox fine, and the buttons are working when clicked. I think it is sending the image fine to FurnitureMenu, im not totally sure though. That is why I think it may be a problem with the mosue listeners in FurnitureMenu. Will test later.